BDOG-447: Basic What's Running Where
Porting the What's Running Where from the old releases app to the
Catalogue. This is a first commit; more will come to address all the
points in the ticket. For instance, filtering is only available via
query params, and there is no link to the page on the menu yet.