BDOG-447: Filtering, mainly by App Name
Profile filter can be done via the query params, but there's another
ticket somewhere to do profiles properly so haven't completed that bit
yet.
The application-name filtering is only via JavaScript, as there are a
few UX behaviours that feel weird with the current way we filter
server-side. I will spend a bit of time trying to form a model for how
we do this nicely...

Screenshot available on the JIRA.