BDOG-2616 - Add initial Create a Repository page in Catalogue
-Added link on Repositories catalogue page
-Created view, following style of Jenkins jobs. Note that a deliberate decision
was made NOT to use the Play field form helpers in the view template. This is
because in order for information to flow semantically correct on the form, for
each field, we would want the label (e.g. repository type), extra information, and
then the form field itself. However if you use the play helpers, they will always
construct a label (even if not specified), meaning that the label will always be
directly before the input field, meaning any extra information for that field
would have to be before the fields heading label, or after the input itself.
-Created controller and form. Form uses the Play verifying/contraints helpers.
-Created an Enum for CreateARepositoryType for easier validation